requirements.txt
coverage==7.6.4
dacite==1.8.1
dependency-injector==4.42.0
Faker==30.8.1
Flask==3.0.3
gcp-microservice-utils==0.4.0
google-cloud-firestore==2.19.0
gunicorn==23.0.0
marshmallow==3.23.0
marshmallow_dataclass==8.7.1
mypy==1.13.0
requests==2.32.3
responses==0.25.3
ruff==0.7.1
tightwrap==24.3.0
types-passlib==1.7.7.20240819
types-requests==2.32.0.20241016
unittest-parametrize==1.5.0
unittest-xml-reporting==3.2.0
